Added Bootstrap into html.index, added photo in pop up for google maps marker, added nav bar + jumbtron
Sorry, I didn't realise all the changes saved to the master branch and
not my own!
I've added bootstrap into the html page, as well as another CSS page for
other styling.
Added a navigation bar and jumbotron, and changed fonts and colour of
them & the background of body.
Changed the photo for iceland pop up and linked the iceland page to it.