Neatened up website
Sorry, had to commit my changes to master because I have no idea whats
happened to my branch! All I've done is added my own Myanmar page,
linked font awesome properly to all the pages so the button show up, and
made the footers span the width of the page. Also put text in the pop up
boxes on the map to make it look neater.